requirements.txt
django-pwa
django==5.1
python-dotenv
pillow
djlint
onnxruntime
rembg
crispy-bootstrap4
django-crispy-forms
xendit-python
python-dateutil
requests
gunicorn
psycopg2-binary
sib-api-v3-sdk
google-generativeai
google-genai